server/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
docopt==0.6.2
grpcio==1.54.2
idna==3.4
protobuf==4.23.2
psycopg2==2.9.6
requests==2.31.0
urllib3==2.0.2
yarg==0.1.9
